requirements.txt
streamlit
google-generativeai
python-dotenv
langchain
faiss-cpu
langchain_google_genai
nltk
pdfplumber
tabula
python-docx